+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added condition Contact Custom Field Changed. Need to test #133.
 * Fixed #133: Contact Custom Data changed trigger works now with pre data. Requires `hook_civicrm_customPre`. Available since CiviCRM version 5.28.
 * Fixed #134: Made the GroupContact trigger compatible with all signatures.
+* Fixed #135: When custom data on a contact changes then also fire the contact is changed trigger.
 * Fixed regression bug from !96 and refactored cleaned up the code.
 
 ## Version 2.23